COSI-72: add-and-use-common-iam-mock
